requirements.txt
flask
inject
flask-lambda
PyJWT
boto3
flask-cors